NR-103816: add fail test case for cloud http client timeout
